backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "artIlluminati",email = "artilluminatigd@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=45.0.5,<46.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiogram (>=3.21.0,<4.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
